+++ Sources/Views/Cells/MessageCollectionViewCell.swift
@@ -131,6 +131,15 @@ open class MessageCollectionViewCell: UICollectionViewCell, CollectionViewReusab
             break
         }
     }
+    
+    /// Handle long press gesture, return true when gestureRecognizer's touch point in `messageContainerView`'s frame
+    open override func gestureRecognizerShouldBegin(_ gestureRecognizer: UIGestureRecognizer) -> Bool {
+        let touchPoint = gestureRecognizer.location(in: self)
+        if gestureRecognizer.isKind(of: UILongPressGestureRecognizer.self) {
+            return messageContainerView.frame.contains(touchPoint)
+        }
+        return false
+    }
 
     /// Handle `ContentView`'s tap gesture, return false when `ContentView` doesn't needs to handle gesture
     open func cellContentView(canHandle touchPoint: CGPoint) -> Bool {